refactor(ZNTA-1835): fix relative imports in editor test files
